+++ src/textual/_compositor.py
@@ -614,6 +614,7 @@ class Compositor:
                             )
                         )
                         widget.set_reactive(Widget.scroll_y, new_scroll_y)
+                        widget.set_reactive(Widget.scroll_target_y, new_scroll_y)
                         widget.vertical_scrollbar._reactive_position = new_scroll_y
 
                     if visible_only: